requirements.txt
streamlit==1.28.0
spacy==3.7.2
nltk==3.8.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
PyPDF2==3.0.1
python-docx==0.8.11
pdfplumber==0.9.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
wordcloud==1.9.2
textstat==0.7.3